chore: Move PropertyKeySet to own file, remove dead code in EnumerableOwnProperties
